+++ src/netmanager-app/components/devices/import-device-form.tsx
@@ -24,7 +24,7 @@ import {
 import { DialogClose } from "@/components/ui/dialog";
 import { useRef } from "react";
 import { useToast } from "@/components/ui/use-toast";
-import { useCreateDevice } from "@/core/hooks/useDevices";
+import { useImportDevice } from "@/core/hooks/useDevices";
 
 const CATEGORIES = [
   { value: 'lowcost', name: 'Lowcost' },
@@ -32,61 +32,67 @@ const CATEGORIES = [
   { value: 'gas', name: 'GAS' }
 ];
 
-const airqoDeviceSchema = z.object({
+const importDeviceSchema = z.object({
   name: z.string().min(2, {
     message: "Device name must be at least 2 characters.",
+  }),
+  deviceNumber: z.string().min(1, {
+    message: "Device number is required.",
   }),
   category: z.string({
     required_error: "Please select a category.",
   }),
   description: z.string().optional(),
   network: z.string(),
+  writeKey: z.string().optional(),
+  readKey: z.string().optional(),
 });
 
-type AirQoDeviceValues = z.infer<typeof airqoDeviceSchema>;
+type ImportDeviceValues = z.infer<typeof importDeviceSchema>;
 
-export function AddAirQoDeviceForm() {
+export function ImportDeviceForm() {
   const { toast } = useToast();
   const cancelButtonRef = useRef<HTMLButtonElement>(null);
-  const { mutate: createDevice, isPending } = useCreateDevice();
+  const { mutate: importDevice, isPending } = useImportDevice();
 
-  const form = useForm<AirQoDeviceValues>({
-    resolver: zodResolver(airqoDeviceSchema),
+  const form = useForm<ImportDeviceValues>({
+    resolver: zodResolver(importDeviceSchema),
     defaultValues: {
       name: "",
+      deviceNumber: "",
       category: "",
       description: "",
       network: "airqo",
+      writeKey: "",
+      readKey: "",
     },
   });
 
-  function onSubmit(values: AirQoDeviceValues) {
-    // Create base data object with required fields
+  function onSubmit(values: ImportDeviceValues) {
     const deviceData = {
       long_name: values.name,
+      device_number: parseInt(values.deviceNumber),
       category: values.category,
       network: values.network,
+      ...(values.description?.trim() && { description: values.description.trim() }),
+      ...(values.writeKey?.trim() && { writeKey: values.writeKey.trim() }),
+      ...(values.readKey?.trim() && { readKey: values.readKey.trim() })
     };
 
-    // Add optional fields only if they have values
-    if (values.description?.trim()) {
-      deviceData['description'] = values.description.trim();
-    }
-
-    createDevice(
+    importDevice(
       deviceData,
       {
         onSuccess: () => {
           toast({
             title: "Success",
-            description: "AirQo device added successfully",
+            description: "Device imported successfully",
           });
           cancelButtonRef.current?.click();
         },
         onError: (error: Error) => {
           toast({
             title: "Error",
-            description: error.message || "Failed to add device",
+            description: error.message || "Failed to import device",
             variant: "destructive",
           });
         },
@@ -106,6 +112,20 @@ export function AddAirQoDeviceForm() {
                 <FormLabel>Device Name *</FormLabel>
                 <FormControl>
                   <Input placeholder="Enter device name" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="deviceNumber"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Device Number *</FormLabel>
+                <FormControl>
+                  <Input placeholder="Enter device number" {...field} />
                 </FormControl>
                 <FormMessage />
               </FormItem>
@@ -153,6 +173,34 @@ export function AddAirQoDeviceForm() {
 
           <FormField
             control={form.control}
+            name="writeKey"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Write Key</FormLabel>
+                <FormControl>
+                  <Input placeholder="Enter write key" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="readKey"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Read Key</FormLabel>
+                <FormControl>
+                  <Input placeholder="Enter read key" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
             name="description"
             render={({ field }) => (
               <FormItem className="col-span-2">
@@ -176,10 +224,10 @@ export function AddAirQoDeviceForm() {
             </Button>
           </DialogClose>
           <Button type="submit" disabled={isPending}>
-            {isPending ? "Adding..." : "Add Device"}
+            {isPending ? "Importing..." : "Import Device"}
           </Button>
         </div>
       </form>
     </Form>
   );
-} 
+}
